+++ inngest/src/result.rs
@@ -11,69 +11,12 @@ use axum::{
 use serde::Serialize;
 use serde_json::{json, Value};
 
-use crate::{
-    handler::{IntrospectResult, SyncResponse},
-    header,
-};
+use crate::header;
 
 #[derive(Serialize)]
 pub struct SdkResponse {
     pub status: u16,
     pub body: Value,
-}
-
-impl IntoResponse for SdkResponse {
-    fn into_response(self) -> axum::response::Response {
-        let mut headers = HeaderMap::new();
-        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
-        headers.insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/json"),
-        );
-        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
-        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
-        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
-
-        match self.status {
-            200 => (StatusCode::OK, headers, Json(self.body)).into_response(),
-            206 => (StatusCode::PARTIAL_CONTENT, headers, Json(self.body)).into_response(),
-            400 => (StatusCode::BAD_REQUEST, headers, Json(self.body)).into_response(),
-            500 => (StatusCode::INTERNAL_SERVER_ERROR, headers, Json(self.body)).into_response(),
-            _ => (StatusCode::BAD_REQUEST, Json(json!("Unknown response"))).into_response(),
-        }
-    }
-}
-
-impl IntoResponse for SyncResponse {
-    fn into_response(self) -> axum::response::Response {
-        let mut headers = HeaderMap::new();
-        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
-        headers.insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/json"),
-        );
-        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
-        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
-        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
-
-        (StatusCode::OK, headers, Json(&self)).into_response()
-    }
-}
-
-impl IntoResponse for IntrospectResult {
-    fn into_response(self) -> axum::response::Response {
-        let mut headers = HeaderMap::new();
-        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
-        headers.insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/json"),
-        );
-        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
-        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
-        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
-
-        (StatusCode::OK, headers, Json(&self)).into_response()
-    }
 }
 
 /// Error type that the user (developer) is supposed to interact with
