+++ inngest/src/serve/axum.rs
@@ -1,15 +1,18 @@
 use crate::{
     basic_error,
     handler::{Handler, IntrospectResult, RunQueryParams, SyncQueryParams, SyncResponse},
-    header::Headers,
+    header::{self, Headers},
     result::{Error, SdkResponse},
 };
 
 use axum::{
     extract::{Query, State},
-    http::HeaderMap,
+    response::IntoResponse,
+    http::{HeaderMap, HeaderValue, StatusCode},
+    Json
 };
 use serde::Deserialize;
+use serde_json::json;
 use std::{fmt::Debug, sync::Arc};
 
 const FRAMEWORK: &str = "axum";
@@ -47,5 +50,61 @@ where
     match serde_json::from_str(&raw) {
         Ok(body) => handler.run(&headers, &query, &raw, &body).await,
         Err(_err) => Err(basic_error!("failed to parse body as JSON")),
+    }
+}
+
+
+// Response conversion
+impl IntoResponse for SdkResponse {
+    fn into_response(self) -> axum::response::Response {
+        let mut headers = HeaderMap::new();
+        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
+        headers.insert(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/json"),
+        );
+        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
+        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
+        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
+
+        match self.status {
+            200 => (StatusCode::OK, headers, Json(self.body)).into_response(),
+            206 => (StatusCode::PARTIAL_CONTENT, headers, Json(self.body)).into_response(),
+            400 => (StatusCode::BAD_REQUEST, headers, Json(self.body)).into_response(),
+            500 => (StatusCode::INTERNAL_SERVER_ERROR, headers, Json(self.body)).into_response(),
+            _ => (StatusCode::BAD_REQUEST, Json(json!("Unknown response"))).into_response(),
+        }
+    }
+}
+
+impl IntoResponse for SyncResponse {
+    fn into_response(self) -> axum::response::Response {
+        let mut headers = HeaderMap::new();
+        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
+        headers.insert(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/json"),
+        );
+        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
+        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
+        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
+
+        (StatusCode::OK, headers, Json(&self)).into_response()
+    }
+}
+
+impl IntoResponse for IntrospectResult {
+    fn into_response(self) -> axum::response::Response {
+        let mut headers = HeaderMap::new();
+        let sdk = format!("rust:{}", env!("CARGO_PKG_VERSION"));
+        headers.insert(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/json"),
+        );
+        headers.insert(header::INNGEST_FRAMEWORK, HeaderValue::from_static("axum"));
+        headers.insert(header::INNGEST_SDK, HeaderValue::from_str(&sdk).unwrap());
+        headers.insert(header::INNGEST_REQ_VERSION, HeaderValue::from_static("1"));
+
+        (StatusCode::OK, headers, Json(&self)).into_response()
     }
 }
